add exception handling
